move useAuth hook to AuthContext
